# Lexer performance testing with new streams