… runtime-testsuite/test/org/antlr/v4/test/runtime/java/api/Java.g4
… runtime-testsuite/test/org/antlr/v4/test/runtime/java/api/perf/emoji.txt
… runtime-testsuite/test/org/antlr/v4/test/runtime/java/api/perf/udhr_hin.txt
… runtime-testsuite/test/org/antlr/v4/test/runtime/java/api/perf/udhr_kor.txt