This PR is an attempt to measure the speed of the new Java-target unicode streams by @bhamiltoncx and lexing with them. The legacy streams, such as `ANTLRFileStream` are tested as a baseline. Each lexer test is run multiple times (1500 or 500) and the highest 20% times are dropped; in Java these are likely  garbage collection or compilation pauses. The numbers seem reasonably stable across multiple runs (i.e., variance is low around that average time).

```
Warming up Java compiler....
   load_legacy_java_ascii average time    53us over 3500 loads of 29038 symbols from Parser.java
    load_legacy_java_utf8 average time    42us over 3500 loads of 29038 symbols from Parser.java
    load_legacy_java_utf8 average time   121us over 3500 loads of 13379 symbols from udhr_hin.txt
            load_new_utf8 average time   193us over 3500 loads of 29038 symbols from Parser.java
            load_new_utf8 average time   199us over 3500 loads of 13379 symbols from udhr_hin.txt

    lex_legacy_java_ascii average time   399us over 2000 runs of 29038 symbols
    lex_legacy_java_ascii average time   918us over 2000 runs of 29038 symbols DFA cleared
     lex_legacy_java_utf8 average time   377us over 2000 runs of 29038 symbols
     lex_legacy_java_utf8 average time   925us over 2000 runs of 29038 symbols DFA cleared
        lex_new_java_utf8 average time   460us over 2000 runs of 29038 symbols
        lex_new_java_utf8 average time   989us over 2000 runs of 29038 symbols DFA cleared

 lex_legacy_grapheme_utf8 average time  6862us over  400 runs of  6614 symbols from udhr_kor.txt
 lex_legacy_grapheme_utf8 average time  7023us over  400 runs of  6614 symbols from udhr_kor.txt DFA cleared
 lex_legacy_grapheme_utf8 average time  6290us over  400 runs of 13379 symbols from udhr_hin.txt
 lex_legacy_grapheme_utf8 average time  6238us over  400 runs of 13379 symbols from udhr_hin.txt DFA cleared
    lex_new_grapheme_utf8 average time  6863us over  400 runs of  6614 symbols from udhr_kor.txt
    lex_new_grapheme_utf8 average time  7014us over  400 runs of  6614 symbols from udhr_kor.txt DFA cleared
    lex_new_grapheme_utf8 average time  6187us over  400 runs of 13379 symbols from udhr_hin.txt
    lex_new_grapheme_utf8 average time  6284us over  400 runs of 13379 symbols from udhr_hin.txt DFA cleared
    lex_new_grapheme_utf8 average time    99us over  400 runs of    85 symbols from emoji.txt
    lex_new_grapheme_utf8 average time   111us over  400 runs of    85 symbols from emoji.txt DFA cleared
```