@bhamiltoncx 2017-03-17
Nice! I'll see if I can get this running under YourKit. Probably some low-hanging fruit in the new streams.