@parrt 2017-03-17
@antlr/antlr-targets This might be a useful benchmark for your targets. Java profiling is notoriously challenging given the garbage collection and compilation pauses that occur at runtime. Other targets may not have this issue.