@parrt 2017-03-17
they do seem pretty close but the new streams are consistently a tiny bit slower during lexing, which doesn't necessarily make sense as you say. hmm... could be an artifact of the way I'm testing