@bhamiltoncx 2017-03-17
Yeah, I see the small difference. If it's consistent, it's probably real! The `CodePointCharStream` methods basically delegate to `IntBuffer`, which seems fairly optimal:

http://grepcode.com/file/repository.grepcode.com/java/root/jdk/openjdk/8u40-b25/java/nio/Buffer.java#Buffer

Maybe the JIT is inlining the methods in `ANTLRInputStream` which makes them a teeny bit faster? Is it worth chasing this down?