@parrt 2017-03-17
I'm noticing that the order in which I test these seems to matter for the loads, despite trying to warm up the compiler. damn this stuff is fiddly. I don't think it's worth chasing this performance down until we see it on a real problem. If there is a tiny advantage people can use the legacy streams.